TGUnit.lua: Add support for spell cast.
This monitors the player's spell cast using events and the
Casting/ChannelingInfo() functions for precise cast tracking.  It also
monitors the *start* of other unit casting by parsing the combat log.
It does not reliably detect the *end* of other unit casting, but this
would be sufficient for a unit frame to "pulse" a spell name for a few
seconds showing that some cast had begun.